+++ lib/objectbox.dart
@@ -11,6 +11,8 @@ export 'src/common.dart';
 export 'src/model.dart';
 export 'src/modelinfo/index.dart';
 export 'src/query/query.dart';
+export 'src/relations/info.dart';
 export 'src/relations/to_one.dart';
+export 'src/relations/to_many.dart' hide InternalToManyTestAccess;
 export 'src/store.dart';
 export 'src/sync.dart';