+++ lib/src/modelinfo/index.dart
@@ -4,3 +4,4 @@ export 'entity_definition.dart';
 export 'model_definition.dart';
 export 'modelinfo.dart';
 export 'modelproperty.dart';
+export 'modelrelation.dart';