+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `ToOne<>` class to wrap related entities. See examples for details.
   (thanks [@Buggaboo](https://github.com/Buggaboo) for jump-starting this).
+* Add `ToMany<>` class to wrap related entities. See examples for details.
 * Significantly improve read and write performance of all box operations.
 * Change `Box.put()` and `putMany()` - now also update given object's ID property.
   Note: If you previously `put()` the same new object instance multiple times without setting id, the object will now